Move the schema-change regression to the sqlite integration suite
The test writes a sqlite database to disk, runs DDL against it, and validates
twice, so pytest.mark.unit was wrong on both counts the project's testing ADR
names: it talks to a database and it touches the file system. It also covers a
behavioral change to a data source and a validation mechanic, which the
contributor guide requires be covered in
tests/integration/data_sources_and_expectations.

It now lives beside the other sqlite data source tests, marked integration and
sqlite, so it runs in the sqlite marker lane. Reverting the inspector reset
still fails it, so it discriminates the same defect it did before the move.
